Only remove the placeholder on the first keyup
Fixes issue where inserting text into the middle of the textbox
would cause the cursor to jump to the end after every keypress
(due to setting the `value` property).

Also helps avoid the case where the valid input CONTAINS the placeholder
text. For example: Placeholder is "123" and valid input is "51234".