Fix reload bug
When doing page reloading (Ctrl+R in IE8, for example), since in
main.js#L102 the element value is set and not cleared during the
reloading. So the second time main.js#L100 will evaluate to false, thus
no placeholder is rendered. Instead, the input will have the place
holder value as the actual value. This commit fixed that.