Fix #75 - Bug with form submit handler not being attached in some cases
According to the HTML5 spec, the 'form' property of an HTMLInputElement
may be any id of a form in the current document *or* the containing form
HTMLFormElement itself. The second case was previously not accounted for,
only adding the submit handler when a string was present. This was
breaking functionality on IE9.

Credit for this goes to @littletijn: http://git.io/vRYQX